requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
instructor>=1.9.0
openai>=1.70.0
pydantic>=2.8.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0 